- sort static files into to their projects - new location for themes html-dist - make build themes script a bit faster and simpler
