+++ .gitignore
@@ -77,9 +77,6 @@ static/conversations.css
 # portable config folders
 DeltaChatData
 
-static/fallback-theme.css
-themes/*.css
-
 # ignore provisioning file
 embedded.provisionprofile
