+++ packages/frontend/.gitignore
@@ -1,0 +1,4 @@
+static/fallback-theme.css
+themes/*.css
+
+html-dist
