+++ packages/frontend/package.json
@@ -5,13 +5,13 @@
   "license": "GPL-3.0-or-later",
   "scripts": {
     "check:types": "tsc --noEmit",
+    "build": "pnpm build:ts && pnpm build:themes && pnpm build:static && pnpm build:source-map-support && pnpm build:scss",
     "build:ts": "node ./bin/build-frontend-ts.mjs",
     "build:source-map-support": "node ./bin/build-frontend-source_map_support.js",
     "build:scss": "sass ./scss/manifest.scss ./html-dist/main.css --embed-source-map --style=compressed",
-    "WIP": "exit 1 # scripts below this still are not done",
-    "build": "pnpm build:ts && pnpm build:themes && pnpm build:static && pnpm build:source-map-support && pnpm build:scss",
+    "build:static": "node ../../bin/copy.js ../../static ./html-dist && node ../../bin/copy.js ./static ./html-dist",
     "build:themes": "sass themes/light.scss static/fallback-theme.css --style=compressed --embed-source-map && node ./bin/build-frontend-themes.js",
-    "build:static": "node ./bin/copy.js ../static ./html-dist",
+    "WIP": "exit 1 # scripts below this still are not done",
     "watch": "pnpm build:source-map-support && pnpm watch:ts & pnpm watch:static & pnpm watch:scss & pnpm watch:types",
     "watch:ts": "pnpm build:ts -- -w",
     "watch:types": "npx tsc -b src/renderer --pretty -w --preserveWatchOutput",
