+++ packages/frontend/static/avoid-eval.js
