+++ packages/frontend/static/fix-missing-vars.js
