+++ packages/frontend/themes/dev_minimal.scss
