+++ packages/target-electron/static/electron_html_email_view/electron_html_email_view.js
